Change timestamps to int, not double

+++ Savour/Deal.swift
@@ -26,7 +26,7 @@ class DealData{
     var endTime: Double?
     var favorited: Bool?
     var redeemed: Bool?
-    var redeemedTime: Double?
+    var redeemedTime: Int?
     var type: String?
     var code: String?
     var activeHours: String?
@@ -96,8 +96,8 @@ class DealData{
             self.id = snap?.key
             self.code = value["code"] as? String ?? ""
             if let redeemValue = value["redeemed"] as? NSDictionary{
-                if let time = redeemValue[ID] as? Double{
-                    if Date().timeIntervalSince1970 - time > 60*60*24*7*2 {
+                if let time = redeemValue[ID] as? Int{
+                    if Int(Date().timeIntervalSince1970) - time > 60*60*24*7*2 {
                         //If redeemed 2 weeks ago, allow user to use deal again - Should be changed in the future
                         let randStr = String.random(length: 10)
                         let ref = Database.database().reference().child("Deals").child(self.id!).child("redeemed")
